backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
google-generativeai==0.8.3
gunicorn==22.0.0
django-filter==24.3
